Refactor: reorganize project folder structure for better maintainability
