+++ src/app/components/city-at-home/city-at-home.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
-import { CitiesService } from '../../cities.service';
-import { DataApiService } from '../../data-api.service';
-import { CountryCodePipe } from '../../country-code.pipe';
-import { TempPipe } from '../../temp.pipe';
+import { CitiesService } from '../../services/cities.service';
+import { DataApiService } from '../../services/data-api.service';
+import { CountryCodePipe } from '../../pipes/country-code.pipe';
+import { TempPipe } from '../../pipes/temp.pipe';
 
 @Component({
   selector: 'app-city-at-home',
