+++ src/app/components/city-card/city-card.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
-import { CitiesService } from '../../cities.service';
-import { CountryCodePipe } from '../../country-code.pipe';
-import { TempPipe } from '../../temp.pipe';
+import { CitiesService } from '../../services/cities.service';
+import { CountryCodePipe } from '../../pipes/country-code.pipe';
+import { TempPipe } from '../../pipes/temp.pipe';
 
 @Component({
   selector: 'app-city-card',
