+++ src/app/pipes/temp.pipe.ts
